+++ Magic/CHANGELOG.md
@@ -5,6 +5,8 @@
 ## Fixes
 
  - Fix flying mob spawns and nether mob spawns in `spawnmobs` example
+ - Fix some forms of configuring an AreaOfEffectCloud
+   The potion effect duration now defaults to the same duration as the cloud tself
 
 ## Changes